fix(hermes): defer .env provisioning to setup success + centralize dotenv parser (Codex review)
- setup.ts now provisions API_SERVER_KEY/API_SERVER_ENABLED LAST, after the
  provider-block rewrite and skill write (the steps that can throw). An aborted
  setup (e.g. --preserve-provider against an existing non-DKG provider) no
  longer leaves a new key behind as a side effect. Added a regression test.
- Centralized the python-dotenv value parser as `parseDotenvValue` in
  @origintrail-official/dkg-core (a dependency of both packages, mocked in
  neither relevant test path). The adapter (key detection) and daemon (bearer
  resolution) now share one implementation, so they cannot drift.
